Add Jira Hide unused bulk edit fields
Most of the time the bulk editing ends up fixing the following
fields only:

- Fix Version
- Components
- Labels

All the others are in the way and make tabbing a longer process
to get to that "Next" button at the bottom.

Guess the other option would be to have a shortcut to click that
"Next" button and leave all fields present.

But still would like to skip from having to see some of the fields
in the first place.