Add Projector Redirect back into timesheet
Having to click two buttons on two pages to see the password manager
log back in to then just land and sit on the dashboard is irksome.

This at least will skip those two interstitial pages and move things
into the login page. If the password manager does what it can do and
fills out the details, you will already be back on the timesheet.

Not sure about those who want to stick around on the actual dashboard.