Disable Workflow trigger when in textarea
Noticed this when entering a comment with a code block that the whole
thing set off in transition at the attempt at "<?php".

Needless, it's not helpful to transition while in the middle of giving
an update, so leave it alone.

On the concerned page, it looks like the text fields are all TEXTAREA
as opposed to having INPUT type fields.