Move out URL switching logic to its own function
Adding a bunch of small functions bloats up things a little, but
at least overall it looks like a bit more readable in terms of
soaking it in. And withouth related functions it also trims down
the levels of indentation, which seems to be one factor in the
whole cognitive complexity checks.